requirement.txt
attrs==19.1.0
boto3==1.9.185
botocore==1.12.185
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dash==1.0.1
dash-core-components==1.0.0
dash-html-components==1.0.0
dash-renderer==1.0.0
dash-table==4.0.1
decorator==4.4.0
docutils==0.14
Flask==1.1.1
Flask-Compress==1.4.0
idna==2.8
imutils==0.5.2
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter-core==4.5.0
MarkupSafe==1.1.1
mkl-fft==1.0.12
mkl-random==1.0.2
nbformat==4.4.0
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.1.0
plotly==3.10.0
pyrsistent==0.15.3
pytesseract==0.2.7
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
retrying==1.3.3
s3transfer==0.2.1
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
Werkzeug==0.15.4
wincertstore==0.2
